Use displayPrompt
